requirements.txt
tensorflow==2.8.0
scikit-learn==1.0.2
opencv-python==4.5.3.20210927
google-cloud-vision==2.5.0
google-cloud-storage==1.42.3
pandas==1.3.5
numpy==1.21.4
matplotlib==3.5.1
pytest==6.2.5